requirements.txt
python-telegram-bot
openai
python-dotenv
aiohttp
sqlalchemy
psycopg2-binary
asyncpg 